Move RepoManager to its own source file
