+++ src/ModelManager.h
@@ -8,8 +8,7 @@
 #include "Network.h"
 #include "types.h"
 #include "Settings.h"
-
-constexpr const char* kModelListUrl = "https://translatelocally.com/models.json";
+#include "RepoManager.h"
 
 namespace translateLocally {
     namespace models {
@@ -110,34 +109,6 @@ struct Model {
 };
 
 Q_DECLARE_METATYPE(Model)
-
-class RepoManager : public QAbstractTableModel {
-    Q_OBJECT
-public:
-    RepoManager(QObject * parent, Settings *);
-    /**
-     * @brief getRepos getsAll currently available repos
-     * @return QStringList of URLs
-     */
-    QStringList getRepos();
-    void insert(QStringList new_model);
-    void remove(int index);
-
-    int rowCount(const QModelIndex &parent = QModelIndex()) const;
-    int columnCount(const QModelIndex &parent = QModelIndex()) const;
-    QVariant data(const QModelIndex &index, int role = Qt::DisplayRole) const;
-    QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const;
-
-    enum RepoColumn {
-        RepoName,
-        URL
-    };
-
-    Q_ENUM(RepoColumn);
-private:
-    Settings * settings_;
-
-};
 
 class ModelManager : public QAbstractTableModel {
         Q_OBJECT
